requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bokeh==2.3.2
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cssselect==1.1.0
dask==2021.5.1
distributed==2021.5.1
fake-useragent==0.1.11
finpie==0.1362
fsspec==2021.5.0
HeapDict==1.0.1
html5lib==1.1
idna==2.10
iexfinance==0.5.0
Jinja2==3.0.1
locket==0.2.1
lxml==4.6.3
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parse==1.19.0
partd==1.2.0
Pillow==8.2.0
psutil==5.8.0
pyee==8.1.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tblib==1.7.0
toolz==0.11.1
tornado==6.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
xlrd==2.0.1
yarl==1.6.3
zict==2.0.0
